WIP docs: don't monkey patch for sphinx >= 4
There's a really long existing bug in sphinx that causes it to generate
cross references in places where it really should not:
  https://github.com/sphinx-doc/sphinx/issues/2549

In docs/source/conf.py there's a monkey patch for this, from line 363
and further.

It looks like a fix for this was implemented in v4:
  https://github.com/sphinx-doc/sphinx/pull/8638/commits

    -$ git tag --contains 918086b5590763663c1627578085e528f1358384
    v4.0.0
    [...]

The function that is replaced while applying the workaround has been
changed recently, causing breakage (TypeError: patched_make_field() got
an unexpected keyword argument 'inliner').

    -$ git tag --contains 4534d2d1a5755c8cbc9ef4327eab7e34a85a7de8
    v4.1.0
    [...]

So, combining this information, it seems that when the major version
number of the sphinx lib being used is >= 4, the monkey patch part
should be skipped.